unified `ExtensionsProvider`
