+++ library/react-connect-kit/src/Extensions/Accounts.tsx
@@ -18,7 +18,7 @@ import type { Account, Sync } from '@w3ux/types'
 import { formatAccountSs58 } from '@w3ux/utils'
 import { useEffect, useState } from 'react'
 import { combineLatest } from 'rxjs'
-import { useExtensions } from '../Extensions'
+import { useExtensions } from './Connect'
 import type {
   ExtensionAccountsContextInterface,
   ExtensionAccountsProviderProps,
