+++ library/react-connect-kit/src/Extensions/Provider.tsx
@@ -1,0 +1,18 @@
+/* @license Copyright 2024 w3ux authors & contributors
+SPDX-License-Identifier: GPL-3.0-only */
+
+import { ExtensionAccountsProvider } from './Accounts'
+import { ExtensionsConnectProvider } from './Connect'
+import type { ExtensionAccountsProviderProps } from './types'
+
+export const ExtensionsProvider = ({
+  ss58,
+  dappName,
+  children,
+}: ExtensionAccountsProviderProps) => (
+  <ExtensionsConnectProvider>
+    <ExtensionAccountsProvider ss58={ss58} dappName={dappName}>
+      {children}
+    </ExtensionAccountsProvider>
+  </ExtensionsConnectProvider>
+)
