+++ library/react-connect-kit/src/index.ts
@@ -1,6 +1,5 @@
 /* @license Copyright 2024 w3ux authors & contributors
 SPDX-License-Identifier: GPL-3.0-only */
 
-export * from './ExtensionAccounts'
 export * from './Extensions'
 export * from './HardwareAccounts'
